Overlap is now implemented.
Manged to get overlap handling to work at an optimal level after spending two days figuring out how to do it.

To achieve this, it would have to be implemented using one method because when using seperate methods, the bubbles that was overlapped in xPos was not overlapped in yPos.

In that method, a nested if statement was used.