Redesigned the generate bubbles code
The code has been redesigned due to a render bubbles function being implemented. This is to keep consistant numbers of bubbles generated despite of being overlapped.

If there is an overlap of the bubbles, it does not appear on the screen and it would not count, therefore, a while loop was used to get the exact numbers of bubbles on the screen based on input.

for example, if there was 5 overlaps and 15 bubbles were selected, it would loop 20 times to achieve 15 bubbles on screen.