Read PlayerScores have been now implemented
game scores have been saved to the user defaults.

However, if the user closes the app and restarts the app, plays the game, the highscores will be overwritten. What needs to be done is to add a HighScoreManager class so that the readScore method can be used accross different functions.